+++ go-tracker-models/src/main/kotlin/go/tracker/models/trainer/medals/RegionsValues.kt
@@ -1,14 +1,16 @@
 package go.tracker.models.trainer.medals
 
+import java.math.BigDecimal
+
 data class RegionsValues(
-    var kanto: Long? = null,
-    var johto: Long? = null,
-    var hoenn: Long? = null,
-    var sinnoh: Long? = null,
-    var unova: Long? = null,
-    var kalos: Long? = null,
-    var alola: Long? = null,
-    var galar: Long? = null,
-    var hisui: Long? = null,
-    var paldea: Long? = null,
+    var kanto: BigDecimal? = null,
+    var johto: BigDecimal? = null,
+    var hoenn: BigDecimal? = null,
+    var sinnoh: BigDecimal? = null,
+    var unova: BigDecimal? = null,
+    var kalos: BigDecimal? = null,
+    var alola: BigDecimal? = null,
+    var galar: BigDecimal? = null,
+    var hisui: BigDecimal? = null,
+    var paldea: BigDecimal? = null,
 )
